+++ src/molecules/DzFormModal/DzFormModal.tsx
@@ -10,6 +10,7 @@ import {
 } from './types/DzFormModalTypes';
 import { inquireFormSteps } from './formSteps/inquireFormSteps';
 import { newsletterFormSteps } from './formSteps/newsletterFormSteps';
+import { cn } from '../../utils/classnames';
 
 export type DzFormModalProps = {
   isOpen: boolean;
@@ -19,6 +20,7 @@ export type DzFormModalProps = {
   successTitle?: string;
   successSubtitle?: string;
   errorTitle?: string;
+  ImgElement: any;
   errorSubtitle?: string;
   onSubmit: (formValues: Record<string, any>) => Promise<any>;
   type: typeof FORM_MODAL_TYPE_NAMES[number];
@@ -28,11 +30,13 @@ export type DzFormModalProps = {
   onChange?: (fieldName: string, value: any) => void;
   onDirty?: () => void;
   LinkElement: any;
-  ImgElement: any;
   image?: {
     src: string;
     alt: string;
-  };
+  } | null;
+  primaryCTA?: {
+    text: string | null;
+  } | null;
 };
 
 export type SubmissionResult = {
@@ -64,6 +68,7 @@ export const DzFormModal = ({
   LinkElement = 'a',
   ImgElement,
   image,
+  primaryCTA,
 }: DzFormModalProps) => {
   const formSteps = FORM_TYPES_TO_STEPS[type];
   const [submittedFormValues, setSubmittedFormValues] = useState<
@@ -118,17 +123,21 @@ export const DzFormModal = ({
   formSteps[0].title = title;
   formSteps[0].primarySubtitle = subtitle;
   formSteps[0].CTAProps.description = termsAndConditions;
+  formSteps[0].CTAProps.text = !primaryCTA?.text
+    ? formSteps[0].CTAProps.text
+    : primaryCTA?.text;
 
   return (
     <DzModalContainer
       isOpen={isOpen}
       onClose={onCloseModal}
       disableBackdrop={disableBackdrop}
-      className={
+      className={cn(
+        image ? 'md:h-[37.5rem]' : '',
         isSubmitSuccessful === false && !disableBackdrop
           ? 'border-[1px] border-red-100 border-opacity-25'
           : ''
-      }
+      )}
     >
       {!image ? (
         <DzForm
@@ -160,7 +169,7 @@ export const DzFormModal = ({
           }
         />
       ) : (
-        <div className="flex flex-col md:flex-row max-w-[926px]">
+        <div className="flex flex-col md:flex-row max-w-[926px] h-full">
           <DzMedia
             type={MEDIA_TYPES.IMAGE}
             ImgElement={ImgElement}